requirements.txt
Django==1.5.2
setuptools>=1.1
pip>=1.4.1
pillow>=2.1.0
dj-database-url==0.2.1
dj-static==0.0.5
gunicorn

South>=0.7.5
Markdown>=2.1
postmarkup>=1.2.0

django-annoying>=0.7.6
django-pure-pagination>=0.2.1
sorl-thumbnail>=11.12
pytils>=0.2.3

django-simple-math-captcha
django-forms-bootstrap>=2.0.3
pinax-theme-bootstrap>=3.0a11
pinax-theme-bootstrap-account>=1.0b2
django-user-accounts>=1.0b14

pybbm>=0.14.2

django-pagination>=1.0.7
django-taggit==0.10a1
django-taggit-autosuggest>=0.2
django-taggit-templatetags>=0.4.6dev
django-templatetag-sugar>=0.1
django-tinymce>=1.5.1
gdata>=2.0.18
psycopg2>=2.5.1
python-dateutil>=2.1

pytz>=2013b
